[Python] Reorganize bindings
