+++ bindings/python/multibody/expose-geometry.cpp
@@ -16,16 +16,21 @@
 // <http://www.gnu.org/licenses/>.
 
 #include "pinocchio/bindings/python/python.hpp"
-#include "pinocchio/bindings/python/frame.hpp"
+#include "pinocchio/bindings/python/multibody/geometry-object.hpp"
+#include "pinocchio/bindings/python/multibody/geometry-model.hpp"
+#include "pinocchio/bindings/python/multibody/geometry-data.hpp"
 
 namespace se3
 {
   namespace python
   {
     
-    void exposeFrame()
+    void exposeGeometry()
     {
-      FramePythonVisitor::expose();
+      GeometryObjectPythonVisitor::expose();
+      CollisionPairPythonVisitor::expose();
+      GeometryModelPythonVisitor::expose();
+      GeometryDataPythonVisitor::expose();
     }
     
   } // namespace python
