+++ bindings/python/multibody/geometry-model.hpp
@@ -22,10 +22,10 @@
 #include <eigenpy/exception.hpp>
 #include <eigenpy/eigenpy.hpp>
 
-#include "pinocchio/bindings/python/eigen_container.hpp"
-#include "pinocchio/bindings/python/handler.hpp"
+#include "pinocchio/bindings/python/utils/eigen_container.hpp"
+#include "pinocchio/bindings/python/utils/handler.hpp"
 #include "pinocchio/multibody/geometry.hpp"
-#include "pinocchio/bindings/python/model.hpp"
+#include "pinocchio/bindings/python/multibody/model.hpp"
 
 namespace se3
 {
