+++ bindings/python/multibody/geometry-object.hpp
