+++ bindings/python/multibody/joint/joints-variant.hpp
@@ -21,7 +21,7 @@
 #include <eigenpy/exception.hpp>
 #include <eigenpy/eigenpy.hpp>
 #include "pinocchio/multibody/joint/joint-variant.hpp"
-#include "pinocchio/bindings/python/joints-models.hpp"
+#include "pinocchio/bindings/python/multibody/joint/joints-models.hpp"
 
 namespace se3
 {
