+++ bindings/python/spatial/explog.hpp
