+++ bindings/python/spatial/inertia.hpp
