+++ bindings/python/spatial/motion.hpp
