+++ bindings/python/spatial/se3.hpp
